disaster-intelligence-platform/backend/requirements.txt
# Disaster Intelligence Platform - Production Dependencies
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic-settings==2.5.2

# Database
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
geoalchemy2==0.15.2
alembic==1.13.3

# Async HTTP
httpx==0.27.2
aiohttp==3.10.8

# Redis Cache
redis==5.1.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0

# ML / Data Science
xgboost==2.1.1
scikit-learn==1.5.2
shap==0.46.0
numpy==1.26.4

# Geospatial
rasterio==1.3.11
shapely==2.0.6
geopandas==1.0.1

# Background Jobs
APScheduler==3.10.4

# Utilities
python-dotenv==1.0.1
python-multipart==0.0.12
orjson==3.10.7
